+++ src/gql/resolvers/elements/Block.php
@@ -25,9 +25,14 @@ class Block extends ElementResolver
         $query = self::prepareElementQuery($source, $arguments, $context, $resolveInfo);
         $blocks = $query instanceof BlockQuery ? $query->all() : $query;
 
-        // Queries for child blocks after mutations were returning no results... memoize the blocks so they are found
-        foreach ($blocks as $block) {
-            $block->useMemoized($blocks);
+        if ($query instanceof BlockQuery && $query->level == 0) {
+            // If we have all blocks, memoize them to avoid database calls for child block queries.
+            // This also allows child block queries after mutations to return results... not sure if
+            // there's some caching going on causing it to otherwise return no child blocks, need to
+            // look into it further.
+            foreach ($blocks as $block) {
+                $block->useMemoized($blocks);
+            }
         }
 
         return GqlHelper::applyDirectives($source, $resolveInfo, $blocks);